chore #96: gathering entity 명 수정
* 기존 gathering2 > gathering
* 기존 gathering > newgathering
