+++ ody_flutter/lib/data/entity/mapper/gatherings_mapper.dart
@@ -1,11 +1,11 @@
 import "package:ody_flutter/data/entity/gathering/gatherings_response.dart";
-import "package:ody_flutter/domain/model/gathering2.dart";
+import "package:ody_flutter/domain/model/gathering.dart";
 
 extension ToDomain on GatheringsResponse {
-  List<Gathering2> toModel() => GatheringsResponse(meetings: meetings)
+  List<Gathering> toModel() => GatheringsResponse(meetings: meetings)
       .meetings
       .map(
-        (meeting) => Gathering2(
+        (meeting) => Gathering(
           id: meeting.id,
           name: meeting.name,
           date: meeting.date,
