+++ ody_flutter/lib/domain/repository/gathering_repository.dart
@@ -1,17 +1,17 @@
 import "package:dio/dio.dart";
 import "package:ody_flutter/data/entity/gathering/enter_gathering_request.dart";
 import "package:ody_flutter/data/entity/gathering/gathering_request.dart";
+import "package:ody_flutter/domain/model/new_gathering.dart";
 import "package:ody_flutter/domain/model/gathering.dart";
-import "package:ody_flutter/domain/model/gathering2.dart";
 import "package:ody_flutter/domain/model/gathering_detail.dart";
 import "package:ody_flutter/domain/model/nudge.dart";
 
 abstract class GatheringRepository {
-  Future<Gathering> createGathering(GatheringRequest request);
+  Future<NewGathering> createGathering(GatheringRequest request);
 
   Future<GatheringDetail> fetchGathering(int id);
 
-  Future<List<Gathering2>> fetchGatherings();
+  Future<List<Gathering>> fetchGatherings();
 
   Future<bool> enterGathering(EnterGatheringRequest request);
