+++ ody_flutter/lib/screens/gatherings/gatherings_screen.dart
@@ -8,8 +8,8 @@ import "package:ody_flutter/assets/images/images.dart";
 import "package:ody_flutter/components/ody_top_bar.dart";
 import "package:ody_flutter/config/routes.dart";
 import "package:ody_flutter/di/di.dart";
-import "package:ody_flutter/domain/model/gathering2.dart";
-import "package:ody_flutter/screens/base/base_screen.dart";
+import "package:ody_flutter/domain/model/gathering.dart";
+import "package:ody_flutter/presentation/base/base_screen.dart";
 import "package:ody_flutter/screens/gatherings/gatherings_view_model.dart";
 
 class GatheringsScreen extends StatefulWidget {
@@ -159,7 +159,7 @@ class _GatheringsScreenState extends State<GatheringsScreen> {
         ),
       );
 
-  Widget _buildGatheringItem(Gathering2 gathering) => GestureDetector(
+  Widget _buildGatheringItem(Gathering gathering) => GestureDetector(
         onTap: () async {
           await Navigator.pushNamed(
             context,
@@ -263,7 +263,7 @@ class _GatheringsScreenState extends State<GatheringsScreen> {
         ),
       );
 
-  Widget _buildExpandedGatheringDetails(Gathering2 gathering) => Column(
+  Widget _buildExpandedGatheringDetails(Gathering gathering) => Column(
         crossAxisAlignment: CrossAxisAlignment.start,
         children: [
           Text(
